Started some cleanup.
