+++ ModemDotNet.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26730.15
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ModemDotNet", "ModemDotNet\ModemDotNet.csproj", "{59CAB737-2C24-4DC2-BA9F-23255CC8477D}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{59CAB737-2C24-4DC2-BA9F-23255CC8477D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{59CAB737-2C24-4DC2-BA9F-23255CC8477D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{59CAB737-2C24-4DC2-BA9F-23255CC8477D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{59CAB737-2C24-4DC2-BA9F-23255CC8477D}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {BFF0C753-C50E-41DE-B9D6-CDE973476ACB}
+	EndGlobalSection
+EndGlobal
