+++ ModemDotNet/ModemDotNet.csproj
@@ -1,0 +1,8 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <RootNamespace>mgsoto.Ports.Serial</RootNamespace>
+  </PropertyGroup>
+
+</Project>
